move LibraryPackageRepository to common
